refactor: moved key / metadata commands into separate files
